+++ sss_backend/sss_backend/models/conjunto/persona.py
@@ -2,7 +2,7 @@ from django.db import models
 
 
 class Persona(models.Model):
-    per_codigo = models.IntegerField(primary_key=True)
+    per_codigo = models.BigAutoField(primary_key=True)
     per_nombre = models.CharField(max_length=100, blank=True, null=True)
     per_documento = models.IntegerField()
     per_telefono = models.IntegerField(blank=True, null=True)
