+++ sss_backend/sss_backend/models/conjunto/torre.py
@@ -2,7 +2,7 @@ from django.db import models
 
 
 class Torre(models.Model):
-    tor_codigo = models.IntegerField(primary_key=True)
+    tor_codigo = models.BigAutoField(primary_key=True)
     tor_nombre = models.CharField(max_length=10, blank=True, null=True)
 
     class Meta:
